requirements.txt
celery>=4.0.0
PyYAML>=5.1
redis>=3.2.0
six>=1.7.0
tornado>=5.1.0,<6.0.0